requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
sqlalchemy==2.0.35 
pydantic-settings==2.5.2 
pytest==8.3.3 
httpx==0.27.2 
pre-commit==3.8.0 
python-dotenv==1.0.1 
psycopg2-binary==2.9.9 
pytest-asyncio==0.24.0